+++ filter/all.go
@@ -23,8 +23,8 @@ var All = []baker.FilterDesc{
 	RegexMatchDesc,
 	ReplaceFieldsDesc,
 	SetStringFromURLDesc,
+	SliceDesc,
 	StringMatchDesc,
 	TimestampDesc,
 	TimestampRangeDesc,
-	TruncateDesc,
 }