+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add the `Truncate` filter [#175](https://github.com/AdRoll/baker/pull/175)
+- Add the `Slice` filter [#175](https://github.com/AdRoll/baker/pull/175)
 
 ### Changed